handle duplicate MetadataContainer gracefully
Now, if an error occurs during save, (network, or something unexpected)
and the created MetadataContainer isn't able to be deleted properly, it's
okay. atom-force will catch the duplicate error on next save, and delete
the container and re-add it.

To make sure this doesn't cause containers to be deleted when they shouldn't,
they're named as `atom-force-${userId}`. I think MavensMate does it this way
too.